src/ContentPage/requirements.txt
autopep8==2.0.2
beautifulsoup4==4.12.2
blinker==1.6.2
bs4==0.0.1
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudevents==1.9.0
cryptography==40.0.2
deprecation==2.1.0
dnspython==2.3.0
firebase-admin==6.1.0
Flask==2.3.2
functions-framework==3.3.0
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-pubsub==2.16.0
google-cloud-secret-manager==2.16.1
google-cloud-storage==2.8.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
grpc-google-iam-v1==0.12.6
grpcio==1.54.0
grpcio-status==1.54.0
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
msgpack==1.0.5
packaging==23.1
proto-plus==1.22.2
protobuf==4.22.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
PyJWT==2.6.0
pymongo==4.3.3
pyparsing==3.0.9
requests==2.29.0
rsa==4.9
six==1.16.0
soupsieve==2.4.1
tomli==2.0.1
uritemplate==4.1.1
urllib3==1.26.15
watchdog==3.0.0
Werkzeug==2.3.3
